requirements.txt
Flask==2.0.3
gunicorn==20.1.0
keras==2.8.0
requests==2.27.1
tensorflow==2.8.0
urllib3==1.26.8
numpy==1.22.2
Pillow==9.0.1
